+++ packages/zenrouter/lib/src/internal/type.dart
@@ -32,7 +32,7 @@ typedef RouteUriParserSync<T extends RouteTarget> = T Function(Uri uri);
 /// - [RouteLayout], the base class for layout implementations.
 typedef RouteLayoutBuilder<T extends RouteUnique> =
     Widget Function(
-      Coordinator coordinator,
+      CoordinatorCore coordinator,
       StackPath<T> path,
       RouteLayout<T>? layout,
     );